log-generator/requirements.txt
fake-factory
Faker
pytz
tzlocal
numpy
kafka-python